# backend/docs/swagger.yaml
definitions:
  handlers.ApproveJobInput:
    properties:
      approve:
        type: boolean
    type: object
...
  handlers.CreateJobInput:
    properties:
      description:
        maxLength: 16384
        type: string
      duration:
        maxLength: 128
        type: string
      experience:
        enum:
        - newgrad
        - junior
        - senior
        - manager
        - internship
        type: string
      jobtype:
        enum:
        - fulltime
        - parttime
        - contract
        - casual
        - internship
        type: string
      location:
        maxLength: 128
        type: string
      maxsalary:
        type: integer
      minsalary:
        type: integer
      name:
        maxLength: 128
        type: string
      open:
        type: boolean
      position:
        maxLength: 128
        type: string
    required:
    - description
    - duration
    - experience
    - jobtype
    - location
    - name
    - position
    type: object
  handlers.EditJobInput:
    properties:
      description:
        maxLength: 16384
        type: string
      duration:
        maxLength: 128
        type: string
      experience:
        enum:
        - newgrad
        - junior
        - senior
        - manager
        - internship
        type: string
      jobtype:
        enum:
        - fulltime
        - parttime
        - contract
        - casual
        - internship
        type: string
      location:
        maxLength: 128
        type: string
      maxsalary:
        type: integer
      minsalary:
        type: integer
      name:
        maxLength: 128
        type: string
      open:
        type: boolean
      position:
        maxLength: 128
        type: string
    type: object
...
  handlers.JobResponse:
    properties:
      approvalStatus:
        type: string
      bannerId:
        type: string
      companyId:
        type: string
      companyName:
        type: string
      createdAt:
        type: string
      description:
        type: string
      duration:
        type: string
      experience:
        type: string
      id:
        type: integer
      isOpen:
        type: boolean
      jobType:
        type: string
      location:
        type: string
      maxSalary:
        type: integer
      minSalary:
        type: integer
      name:
        type: string
      photoId:
        type: string
      position:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a sample API for KU-Work
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KU-Work API
  version: "1.0"
paths:
...
  /jobs:
    get:
      description: Retrieves a list of job postings with extensive filtering options.
        Behavior changes based on user role. Companies see their own jobs with application
        stats. Admins can see all jobs. Others see only open, approved jobs.
      parameters:
      - default: 32
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Search keyword for name and description
        in: query
        name: keyword
        type: string
      - collectionFormat: csv
        description: Filter by job type(s)
        in: query
        items:
          type: string
        name: jobtype
        type: array
      - collectionFormat: csv
        description: Filter by experience level(s)
        in: query
        items:
          type: string
        name: experience
        type: array
      - description: Minimum salary filter
        in: query
        name: minsalary
        type: integer
      - description: Maximum salary filter
        in: query
        name: maxsalary
        type: integer
      - description: Filter by open status (company only)
        in: query
        name: open
        type: boolean
      - description: Filter by company ID
        in: query
        name: companyId
        type: string
      - description: Filter by specific job ID
        in: query
        name: id
        type: integer
      - description: Filter by approval status (admin/company only)
        in: query
        name: approvalStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of jobs for a non-company user
          schema:
            properties:
              jobs:
                items:
                  $ref: '#/definitions/handlers.JobResponse'
                type: array
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Fetch job listings
      tags:
      - Jobs
    post:
      consumes:
      - application/json
      description: Allows an authenticated company to create a new job posting. The
        job will be pending approval by an admin.
      parameters:
      - description: Job creation data
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateJobInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created job listing
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new job listing
      tags:
      - Jobs
  /jobs/{id}:
    get:
      description: Retrieves the detailed information for a single job posting by
        its ID.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Job details retrieved successfully
          schema:
            $ref: '#/definitions/handlers.JobResponse'
        "400":
          description: 'Bad Request: Invalid job ID'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get job details
      tags:
      - Jobs
    patch:
      consumes:
      - application/json
      description: Allows a company to edit one of their own job postings. Supports
        partial updates.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Job update data
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/handlers.EditJobInput'
      produces:
      - application/json
      responses:
        "200":
          description: Job updated successfully
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit a job listing
      tags:
      - Jobs
...
  /jobs/{id}/approval:
    post:
      consumes:
      - application/json
      description: Allows an admin to approve or reject a job posting submitted by
        a company.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Approval action
        in: body
        name: approval
        required: true
        schema:
          $ref: '#/definitions/handlers.ApproveJobInput'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve or reject a job listing (Admin only)
      tags:
      - Jobs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"