# backend/handlers/job.go
package handlers

import (
	"fmt"
	"ku-work/backend/helper"
	"ku-work/backend/model"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type JobHandlers struct {
	DB           *gorm.DB
	FileHandlers *FileHandlers
}

func NewJobHandlers(db *gorm.DB) *JobHandlers {
	return &JobHandlers{
		DB:           db,
		FileHandlers: NewFileHandlers(db),
	}
}

// CreateJobInput defines the request body for creating a new job.
type CreateJobInput struct {
	Name        string `json:"name" binding:"required,max=128"`
	Position    string `json:"position" binding:"required,max=128"`
	Duration    string `json:"duration" binding:"required,max=128"`
	Description string `json:"description" binding:"required,max=16384"`
	Location    string `json:"location" binding:"required,max=128"`
	JobType     string `json:"jobtype" binding:"required,oneof='fulltime' 'parttime' 'contract' 'casual' 'internship'"`
	Experience  string `json:"experience" binding:"required,oneof='newgrad' 'junior' 'senior' 'manager' 'internship'"`
	MinSalary   uint   `json:"minsalary"`
	MaxSalary   uint   `json:"maxsalary"`
	Open        bool   `json:"open"`
}

// EditJobInput defines the request body for editing an existing job.
type EditJobInput struct {
	Name        *string `json:"name" binding:"omitempty,max=128"`
	Position    *string `json:"position" binding:"omitempty,max=128"`
	Duration    *string `json:"duration" binding:"omitempty,max=128"`
	Description *string `json:"description" binding:"omitempty,max=16384"`
	Location    *string `json:"location" binding:"omitempty,max=128"`
	JobType     *string `json:"jobtype" binding:"omitempty,oneof='fulltime' 'parttime' 'contract' 'casual' 'internship'"`
	Experience  *string `json:"experience" binding:"omitempty,oneof='newgrad' 'junior' 'senior' 'manager' 'internship'"`
	MinSalary   *uint   `json:"minsalary" binding:"omitempty"`
	MaxSalary   *uint   `json:"maxsalary" binding:"omitempty"`
	Open        *bool   `json:"open" binding:"omitempty"`
}

// ApproveJobInput defines the request body for approving a job.
type ApproveJobInput struct {
	Approve bool `json:"approve"`
}

// JobResponse defines the structure for a single job listing in API responses.
type JobResponse struct {
	ID             uint      `json:"id"`
	CreatedAt      time.Time `json:"createdAt"`
	UpdatedAt      time.Time `json:"updatedAt"`
	Name           string    `json:"name"`
	CompanyID      string    `json:"companyId"`
	PhotoID        string    `json:"photoId"`
	BannerID       string    `json:"bannerId"`
	CompanyName    string    `json:"companyName"`
	Position       string    `json:"position"`
	Duration       string    `json:"duration"`
	Description    string    `json:"description"`
	Location       string    `json:"location"`
	JobType        string    `json:"jobType"`
	Experience     string    `json:"experience"`
	MinSalary      uint      `json:"minSalary"`
	MaxSalary      uint      `json:"maxSalary"`
	ApprovalStatus string    `json:"approvalStatus"`
	IsOpen         bool      `json:"open"`
}

// JobWithStatsResponse extends JobResponse with application statistics for company users.
type JobWithStatsResponse struct {
	JobResponse
	Pending  int64 `json:"pending"`
	Accepted int64 `json:"accepted"`
	Rejected int64 `json:"rejected"`
}

// @Summary Create a new job listing
// @Description Allows an authenticated company to create a new job posting. The job will be pending approval by an admin.
// @Tags Jobs
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param job body handlers.CreateJobInput true "Job creation data"
// @Success 200 {object} object{id=uint} "Successfully created job listing"
// @Failure 400 {object} object{error=string} "Bad Request"
// @Failure 401 {object} object{error=string} "Unauthorized"
// @Failure 500 {object} object{error=string} "Internal Server Error"
// @Router /jobs [post]
func (h *JobHandlers) CreateJobHandler(ctx *gin.Context) {
	// Get user ID from context (auth middleware)
	probUserId, hasUserId := ctx.Get("userID")

	// Return error if user ID is not found
	if !hasUserId {
		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
		return
	}
	userid := probUserId.(string)

	input := CreateJobInput{}
	err := ctx.Bind(&input)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Validate input data
	if input.MaxSalary < input.MinSalary {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "minsalary must be lower than or equal to maxsalary"})
		return
	}
	company := model.Company{
		UserID: userid,
	}
	if result := h.DB.First(&company); result.Error != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
		return
	}
	job := model.Job{
		Name:           input.Name,
		CompanyID:      company.UserID,
		Position:       input.Position,
		Duration:       input.Duration,
		Description:    input.Description,
		Location:       input.Location,
		JobType:        model.JobType(input.JobType),
		Experience:     model.ExperienceType(input.Experience),
		MinSalary:      input.MinSalary,
		MaxSalary:      input.MaxSalary,
		ApprovalStatus: model.JobApprovalPending,
		IsOpen:         input.Open,
	}

	// Create Job into database
	if result := h.DB.Create(&job); result.Error != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
		return
	}

	// return job ID on success
	ctx.JSON(http.StatusOK, gin.H{
		"id": job.ID,
	})
}

// @Summary Fetch job listings
// @Description Retrieves a list of job postings with extensive filtering options. Behavior changes based on user role. Companies see their own jobs with application stats. Admins can see all jobs. Others see only open, approved jobs.
// @Tags Jobs
// @Security BearerAuth
// @Produce json
// @Param limit query uint false "Pagination limit" default(32)
// @Param offset query uint false "Pagination offset"
// @Param location query string false "Filter by location"
// @Param keyword query string false "Search keyword for name and description"
// @Param jobtype query []string false "Filter by job type(s)"
// @Param experience query []string false "Filter by experience level(s)"
// @Param minsalary query uint false "Minimum salary filter"
// @Param maxsalary query uint false "Maximum salary filter"
// @Param open query bool false "Filter by open status (company only)"
// @Param companyId query string false "Filter by company ID"
// @Param id query uint false "Filter by specific job ID"
// @Param approvalStatus query string false "Filter by approval status (admin/company only)"
// @Success 200 {object} object{jobs=[]JobWithStatsResponse} "List of jobs for a company user (includes stats)"
// @Success 200 {object} object{jobs=[]JobResponse} "List of jobs for a non-company user"
// @Failure 400 {object} object{error=string} "Bad Request"
// @Failure 500 {object} object{error=string} "Internal Server Error"
// @Router /jobs [get]
func (h *JobHandlers) FetchJobsHandler(ctx *gin.Context) {
	userId := ctx.MustGet("userID").(string)
	// List of query parameters for filtering jobs
	type FetchJobsInput struct {
		Limit          uint     `json:"limit" form:"limit" binding:"max=128"`
		Offset         uint     `json:"offset" form:"offset"`
		Location       string   `json:"location" form:"location" binding:"max=128"`
		Keyword        string   `json:"keyword" form:"keyword" binding:"max=256"`
		JobType        []string `json:"jobtype" form:"jobtype" binding:"max=5,dive,max=32"`
		Experience     []string `json:"experience" form:"experience" binding:"max=5,dive,max=32"`
		MinSalary      uint     `json:"minsalary" form:"minsalary"`
		MaxSalary      uint     `json:"maxsalary" form:"maxsalary"`
		Open           *bool    `json:"open" form:"open"`
		CompanyID      string   `json:"companyId" form:"companyId" binding:"max=64"`
		JobID          *uint    `json:"id" form:"id" binding:"omitempty,max=64"`
		ApprovalStatus string   `json:"approvalStatus" form:"approvalStatus" binding:"max=64"`
	}

	// Set default values for some fields and bind the input
	input := FetchJobsInput{
		MinSalary: 0,
		MaxSalary: ^uint(0) >> 1,
		Limit:     32,
		Offset:    0,
	}
	err := ctx.ShouldBind(&input)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	role := helper.GetRole(userId, h.DB)

	query := h.DB.Model(&model.Job{})

	// Optional id limit
	if input.JobID != nil {
		query = query.Where(&model.Job{
			ID: *input.JobID,
		})
	}

	// Filter Job post by keyword
	if input.Keyword != "" {
		keywordPattern := fmt.Sprintf("%%%s%%", input.Keyword)
		query = query.Where(h.DB.Where("name ILIKE ?", keywordPattern).Or("description ILIKE ?", keywordPattern))
	}

	// Filter Job post by salary range
	query = query.Where("min_salary >= ?", input.MinSalary)
	query = query.Where("max_salary <= ?", input.MaxSalary)

	// Company should only see their own jobs
	if role == helper.Company {
		query = query.Where("company_id = ?", userId)
		if input.Open != nil {
			query = query.Where("is_open = ?", *input.Open)
		}
	} else {
		// Non-company users can filter by company ID if provided
		if input.CompanyID != "" {
			query = query.Where("company_id = ?", input.CompanyID)
		}
		query = query.Where("is_open = ?", true)
	}

	// Filter Job post by location
	if len(input.Location) != 0 {
		query = query.Where("location = ?", input.Location)
	}

	// Filter Job post by job type
	if len(input.JobType) != 0 {
		query = query.Where("job_type IN ?", input.JobType)
	}

	// Filter Job post by experience
	if len(input.Experience) != 0 {
		query = query.Where("experience IN ?", input.Experience)
	}

	// Only Admin and Company can see unapproved jobs
	if role == helper.Admin || role == helper.Company {
		// If is admin, or company then consider approval status
		if input.ApprovalStatus != "" {
			query = query.Where(&model.Job{ApprovalStatus: model.JobApprovalStatus(input.ApprovalStatus)})
		}
	} else {
		// Non-admin and non-company users can only see approved jobs
		query = query.Where(&model.Job{ApprovalStatus: model.JobApprovalAccepted})
	}

	// Offset and Limit
	query = query.Offset(int(input.Offset)).
		Limit(int(input.Limit)).
		Joins("INNER JOIN users ON users.id = jobs.company_id").
		Joins("INNER JOIN companies ON companies.user_id = jobs.company_id")

	// return Job posts with application statistics
	if role == helper.Company {
		var jobsWithStats []JobWithStatsResponse
		result := query.
			Select("jobs.*, users.username as company_name, companies.photo_id, companies.banner_id, COUNT(CASE WHEN job_applications.status = 'pending' THEN 1 END) AS pending, COUNT(CASE WHEN job_applications.status = 'accepted' THEN 1 END) AS accepted, COUNT(CASE WHEN job_applications.status = 'rejected' THEN 1 END) AS rejected").
			Joins("LEFT JOIN job_applications ON job_applications.job_id = jobs.id").
			Group("jobs.id, users.username, companies.photo_id, companies.banner_id").
			Find(&jobsWithStats)

		if result.Error != nil {
			ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
			return
		}
		ctx.JSON(http.StatusOK, gin.H{
			"jobs": jobsWithStats,
		})
		return
	}

	// return Job posts with company info if not company
	var jobs []JobResponse
	result := query.Select("jobs.*, users.username as company_name, companies.photo_id, companies.banner_id").Find(&jobs)
	if result.Error != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
		return
	}
	ctx.JSON(http.StatusOK, gin.H{
		"jobs": jobs,
	})
}

// @Summary Edit a job listing
// @Description Allows a company to edit one of their own job postings. Supports partial updates.
// @Tags Jobs
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path uint true "Job ID"
// @Param job body handlers.EditJobInput true "Job update data"
// @Success 200 {object} object{message=string} "Job updated successfully"
// @Failure 400 {object} object{error=string} "Bad Request"
// @Failure 401 {object} object{error=string} "Unauthorized"
// @Failure 403 {object} object{error=string} "Forbidden"
// @Failure 404 {object} object{error=string} "Not Found"
// @Failure 500 {object} object{error=string} "Internal Server Error"
// @Router /jobs/{id} [patch]
func (h *JobHandlers) EditJobHandler(ctx *gin.Context) {
	// Get user id from context (auth middleware)
	probUserId, hasUserId := ctx.Get("userID")
	// Denied access if user id is not found
	if !hasUserId {
		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
		return
	}
	userid := probUserId.(string)

	// Convert job id to uint
	jobIdStr := ctx.Param("id")
	jobId64, err := strconv.ParseUint(jobIdStr, 10, 64)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid job id"})
		return
	}
	jobId := uint(jobId64)

	// Get job post from database
	job := &model.Job{
		ID: jobId,
	}
	result := h.DB.First(&job)
	if result.Error != nil {
		if result.Error == gorm.ErrRecordNotFound {
			ctx.JSON(http.StatusNotFound, gin.H{"error": "job not found"})
			return
		}
		ctx.JSON(http.StatusNotFound, gin.H{"error": result.Error.Error()})
		return
	}

	var input EditJobInput
	if err := ctx.ShouldBindJSON(&input); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Denied access if user is not the owner of job post
	if job.CompanyID != userid {
		ctx.JSON(http.StatusForbidden, gin.H{"error": "forbidden"})
		return
	}

	// Update job post with new data
	if input.Name != nil {
		job.Name = *input.Name
	}
	if input.Position != nil {
		job.Position = *input.Position
	}
	if input.Duration != nil {
		job.Duration = *input.Duration
	}
	if input.Description != nil {
		job.Description = *input.Description
	}
	if input.Location != nil {
		job.Location = *input.Location
	}
	if input.JobType != nil {
		job.JobType = model.JobType(*input.JobType)
	}
	if input.Open != nil {
		job.IsOpen = *input.Open
	}
	if input.Experience != nil {
		job.Experience = model.ExperienceType(*input.Experience)
	}
	if input.MinSalary != nil {
		job.MinSalary = *input.MinSalary
	}
	if input.MaxSalary != nil {
		job.MaxSalary = *input.MaxSalary
	}
	if job.MinSalary > job.MaxSalary {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "minsalary cannot exceed maxsalary"})
		return
	}

	result = h.DB.Save(&job)
	if result.Error != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
		return
	}
	ctx.JSON(http.StatusOK, gin.H{"message": "job updated successfully"})
}

// @Summary Approve or reject a job listing (Admin only)
// @Description Allows an admin to approve or reject a job posting submitted by a company.
// @Tags Jobs
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path uint true "Job ID"
// @Param approval body handlers.ApproveJobInput true "Approval action"
// @Success 200 {object} object{message=string} "ok"
// @Failure 400 {object} object{error=string} "Bad Request"
// @Failure 404 {object} object{error=string} "Not Found"
// @Failure 500 {object} object{error=string} "Internal Server Error"
// @Router /jobs/{id}/approval [post]
func (h *JobHandlers) JobApprovalHandler(ctx *gin.Context) {
	input := ApproveJobInput{}
	err := ctx.Bind(&input)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Get job ID from URL parameter
	jobID := ctx.Param("id")

	job := model.Job{}
	result := h.DB.First(&job, "id = ?", jobID)
	if result.Error != nil {
		ctx.JSON(http.StatusNotFound, gin.H{"error": "Job not found"})
		return
	}
	if input.Approve {
		job.ApprovalStatus = model.JobApprovalAccepted
	} else {
		job.ApprovalStatus = model.JobApprovalRejected
	}
	result = h.DB.Save(&job)
	if result.Error != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
		return
	}
	ctx.JSON(http.StatusOK, gin.H{
		"message": "ok",
	})
}

// @Summary Get job details
// @Description Retrieves the detailed information for a single job posting by its ID.
// @Tags Jobs
// @Security BearerAuth
// @Produce json
// @Param id path uint true "Job ID"
// @Success 200 {object} handlers.JobResponse "Job details retrieved successfully"
// @Failure 400 {object} object{error=string} "Bad Request: Invalid job ID"
// @Failure 500 {object} object{error=string} "Internal Server Error"
// @Router /jobs/{id} [get]
func (h *JobHandlers) GetJobDetailHandler(ctx *gin.Context) {
	jobIdStr := ctx.Param("id")
	jobId64, err := strconv.ParseInt(jobIdStr, 10, 64)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid job ID"})
		return
	}
	jobId := uint(jobId64)

	var job JobResponse
	if err := h.DB.Model(&model.Job{}).
		Select("jobs.*, users.username as company_name, companies.photo_id, companies.banner_id").
		Joins("INNER JOIN users ON users.id = jobs.company_id").
		Joins("INNER JOIN companies ON companies.user_id = jobs.company_id").
		First(&job, jobId).Error; err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	ctx.JSON(http.StatusOK, job)
}